Fixed: Invalid Expression
I fixed your invalid expression. The problem was that the call was only needed to be made once at the end and also everything needed to be loaded back to zero after the call. I was thinking of a different way of doing it but this works. Now all we have left is input and output overflow.